@@ -8,41 +8,31 @@
     "build:prod": "ng build --prod",
     "test": "ng test",
     "lint": "ng lint",
-    "mock-apis": "json-server mocks/db.js --port 8585",
-    "doc:serve": "npx compodoc -p tsconfig.app.json -w -s",
-    "doc:build": "npx compodoc -p tsconfig.app.json"
+    "mock-apis": "json-server mocks/db.js --port 8585"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.14",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
-    "@angular/forms": "~8.2.14",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
-    "rxjs": "~6.5.3",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.9.1"
+    "@angular/animations": "~11.0.0",
+    "@angular/common": "~11.0.0",
+    "@angular/compiler": "~11.0.0",
+    "@angular/core": "~11.0.0",
+    "@angular/forms": "~11.0.0",
+    "@angular/platform-browser": "~11.0.0",
+    "@angular/platform-browser-dynamic": "~11.0.0",
+    "@angular/router": "~11.0.0",
+    "rxjs": "~6.6.3",
+    "tslib": "^2.0.3",
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.19",
-    "@angular/cli": "~8.3.19",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
+    "@angular-devkit/build-angular": "~0.1100.0",
+    "@angular/cli": "~11.0.0",
+    "@angular/compiler-cli": "~11.0.0",
     "@compodoc/compodoc": "^1.1.11",
-    "@types/jasmine": "^3.4.6",
-    "@types/node": "~12.12.9",
-    "codelyzer": "^5.0.0",
-    "json-server": "^0.15.1",
-    "karma": "^4.4.1",
-    "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage-istanbul-reporter": "^2.1.0",
-    "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
-    "ts-node": "~8.5.2",
-    "tslint": "~5.20.1",
-    "typescript": "~3.5.3"
+    "@types/node": "~14.14.7",
+    "json-server": "^0.16.2",
+    "ts-node": "~9.0.0",
+    "tslint": "~6.1.3",
+    "typescript": "~4.0.5"
   }
 }
