@@ -1,17 +1,13 @@
-import { RouterModule, Routes } from '@angular/router';
 import { ModuleWithProviders } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
 import { DetailComponent } from './detail.component';
 
 const detailRoutes: Routes = [
   {
-    path     : ':id',
+    path: ':id',
     component: DetailComponent
-  },
-  /*{
-    path: 'detail/:id',
-    component: DetailComponent
-  },*/
+  }
 ];
 
-export const DetailRoute: ModuleWithProviders = RouterModule.forChild(detailRoutes);
+export const DetailRoute: ModuleWithProviders<RouterModule> = RouterModule.forChild(detailRoutes);
 
