@@ -1,17 +1,16 @@
-import { RouterModule, Routes } from '@angular/router';
 import { ModuleWithProviders } from '@angular/core';
-
+import { RouterModule, Routes } from '@angular/router';
 import { ListComponent } from './list.component';
 
 const listRoutes: Routes = [
   {
-    path     : '',
+    path: '',
     component: ListComponent
   },
   {
-    path     : ':category',
+    path: ':category',
     component: ListComponent
-  },
+  }
 ];
 
-export const ListRoute: ModuleWithProviders = RouterModule.forChild(listRoutes);
+export const ListRoute: ModuleWithProviders<RouterModule> = RouterModule.forChild(listRoutes);
