@@ -1,17 +1,12 @@
-import { RouterModule, Routes } from '@angular/router';
 import { ModuleWithProviders } from '@angular/core';
-
+import { RouterModule, Routes } from '@angular/router';
 import { SuggestComponent } from './suggest.component';
 
 const suggestRoutes: Routes = [
   {
-    path     : '',
+    path: '',
     component: SuggestComponent
-  },
-  /*{
-    path: 'suggest',
-    component: SuggestComponent
-  }*/
+  }
 ];
 
-export const SuggestRoute: ModuleWithProviders = RouterModule.forChild(suggestRoutes);
+export const SuggestRoute: ModuleWithProviders<RouterModule> = RouterModule.forChild(suggestRoutes);
