@@ -1,10 +1,8 @@
 import { Component } from '@angular/core';
 
-
 @Component({
-  selector   : 'comics-root',
+  selector: 'comics-root',
   templateUrl: './app.component.html',
-  styleUrls  : ['./app.component.scss']
+  styleUrls: [ './app.component.scss' ]
 })
-export class AppComponent {
-}
+export class AppComponent { }
