@@ -1,11 +1,11 @@
 export const categoriesMock = [
-  {name: 'largo winch', id: 31638},
-  {name: 'asterix', id: 74147},
-  {name: 'xiii', id: 26986},
-  {name: 'gaston', id: 116455},
-  {name: 'tintin', id: 32164},
-  {name: 'thorgal', id: 40719},
-  {name: 'spirou', id: 27780},
-  {name: 'blueberry', id: 40586},
-  {name: 'lucky luke', id: 32180}
+  { name: 'largo winch', id: 31638 },
+  { name: 'asterix', id: 74147 },
+  { name: 'xiii', id: 26986 },
+  { name: 'gaston', id: 116455 },
+  { name: 'tintin', id: 32164 },
+  { name: 'thorgal', id: 40719 },
+  { name: 'spirou', id: 27780 },
+  { name: 'blueberry', id: 40586 },
+  { name: 'lucky luke', id: 32180 }
 ];
