@@ -1,9 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Observable, of } from 'rxjs';
-
 import { CoreServiceModule } from '../core-service.module';
-import { CategoryModel } from './category.model';
 import { categoriesMock } from './categories.data';
+import { CategoryModel } from './category.model';
 
 @Injectable({
   providedIn: CoreServiceModule
