@@ -1,11 +1,4 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 
-@NgModule({
-  declarations: [],
-  imports     : [
-    CommonModule
-  ]
-})
-export class CoreServiceModule {
-}
+@NgModule()
+export class CoreServiceModule { }
