@@ -1,24 +1,23 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SidenavComponent } from './sidenav/sidenav.component';
-import { HeaderComponent } from './header/header.component';
-import { CoreServiceModule } from './core-service.module';
+import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
+import { CoreServiceModule } from './core-service.module';
+import { HeaderComponent } from './header/header.component';
+import { SidenavComponent } from './sidenav/sidenav.component';
 
 @NgModule({
-  declarations: [
-    SidenavComponent,
-    HeaderComponent
-  ],
-  imports     : [
+  imports: [
     CommonModule,
     CoreServiceModule,
     RouterModule
   ],
-  exports     : [
+  declarations: [
+    SidenavComponent,
+    HeaderComponent
+  ],
+  exports: [
     SidenavComponent,
     HeaderComponent
   ]
 })
-export class CoreModule {
-}
+export class CoreModule { }
