@@ -1,5 +1,4 @@
 import { Injectable } from '@angular/core';
-
 import { CoreServiceModule } from '../core-service.module';
 import { UserModel } from './user.model';
 
@@ -11,7 +10,7 @@ export class UserService {
   get connectedUser(): UserModel {
     return {
       firstName: 'Astérix',
-      lastName : 'Le Gaulois'
+      lastName: 'Le Gaulois'
     };
   }
 
