@@ -1,14 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector   : 'comics-detail',
+  selector: 'comics-detail',
   templateUrl: './detail.component.html',
-  styleUrls  : ['./detail.component.scss']
+  styleUrls: [ './detail.component.scss' ]
 })
 export class DetailComponent implements OnInit {
 
-  constructor() {
-  }
+  constructor() { }
 
   ngOnInit() {
   }
