@@ -1,25 +1,24 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ListComponent } from './list.component';
-import { ComicTileComponent } from './comic-tile/comic-tile.component';
-import { SharedModule } from '../shared/shared.module';
+import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
+import { SharedModule } from '../shared/shared.module';
+import { ComicTileComponent } from './comic-tile/comic-tile.component';
+import { ListComponent } from './list.component';
 import { ListRoute } from './list.routing';
 
 @NgModule({
-  declarations: [
-    ListComponent,
-    ComicTileComponent
-  ],
-  imports     : [
+  imports: [
     CommonModule,
     ReactiveFormsModule,
     SharedModule,
     ListRoute
   ],
-  exports     : [
+  declarations: [
+    ListComponent,
+    ComicTileComponent
+  ],
+  exports: [
     ListComponent
   ]
 })
-export class ListModule {
-}
+export class ListModule { }
