@@ -3,96 +3,95 @@ import { ComicModel } from './comic.model';
 // tslint:disable
 export const comicsMock: Array<ComicModel> = [
   {
-    date       : '1961-07-31',
-    id         : 456485,
-    name       : 'Astérix le Gaulois',
-    volumeId   : 74147,
-    imageUrl   : '3897694-1.jpg',
-    description:
-      'Dans le camp fortifié romain Petibonum, on se pose des questions : comment les Irréductibles Gaulois du village d’Astérix font-ils pour ridiculiser encore et toujours la puissance romaine ? Décidé à percer à jour le mystère de la force surhumaine de nos héros, le centurion Caius Bonus envoie un espion déguisé en Gaulois.C’est Caligula Minus qui s’y colle, et découvre bien vite l’existence de la potion magique préparée par Panoramix. Sans attendre, le centurion Caius Bonus fait enlever le druide pour s’emparer de la recette du fameux breuvage histoire, qui sait, de devenir César à la place de Jules ! Une première occasion pour Astérix de démontrer son courage et ses talents de stratège.'
-  },
-  {
-    date       : '1962-07-31',
-    id         : 456486,
-    name       : "La Serpe d'or",
-    volumeId   : 74147,
-    description:
-      'Asterix and Obelix help the Panoramix the druid get a new sickle.'
-  },
-  {
-    date    : '1963-07-31',
-    id      : 458400,
-    name    : 'Astérix et les Goths',
+    date: '1961-07-31',
+    id: 456485,
+    name: 'Astérix le Gaulois',
     volumeId: 74147,
-    imageUrl:
-      '3923229-ast%C3%A9rix%20-%20t03%20-%20ast%C3%A9rix%20et%20les%20goths%20-%20page%201.jpg'
+    imageUrl: '3897694-1.jpg',
+    description: 'Dans le camp fortifié romain Petibonum, on se pose des questions : comment les Irréductibles Gaulois du village d’Astérix font-ils pour ridiculiser encore et toujours la puissance romaine ? Décidé à percer à jour le mystère de la force surhumaine de nos héros, le centurion Caius Bonus envoie un espion déguisé en Gaulois.C’est Caligula Minus qui s’y colle, et découvre bien vite l’existence de la potion magique préparée par Panoramix. Sans attendre, le centurion Caius Bonus fait enlever le druide pour s’emparer de la recette du fameux breuvage histoire, qui sait, de devenir César à la place de Jules ! Une première occasion pour Astérix de démontrer son courage et ses talents de stratège.'
   },
   {
-    date    : '1964-07-31',
-    id      : 458401,
-    name    : 'Astérix Gladiateur',
-    volumeId: 74147
+    date: '1962-07-31',
+    id: 456486,
+    name: "La Serpe d'or",
+    volumeId: 74147,
+    imageUrl: null,
+    description: 'Asterix and Obelix help the Panoramix the druid get a new sickle.'
   },
   {
-    date    : '1965-07-31',
-    id      : 506774,
-    name    : 'Astérix et Cléopâtre',
+    date: '1963-07-31',
+    id: 458400,
+    name: 'Astérix et les Goths',
+    volumeId: 74147,
+    imageUrl: '3923229-ast%C3%A9rix%20-%20t03%20-%20ast%C3%A9rix%20et%20les%20goths%20-%20page%201.jpg'
+  },
+  {
+    date: '1964-07-31',
+    id: 458401,
+    name: 'Astérix Gladiateur',
+    volumeId: 74147,
+    imageUrl: null
+  },
+  {
+    date: '1965-07-31',
+    id: 506774,
+    name: 'Astérix et Cléopâtre',
     volumeId: 74147,
     imageUrl: '4925966-5723389690-81-Rm.jpg'
   },
   {
-    date    : '1966-01-31',
-    id      : 506775,
-    name    : 'Le Combat des chefs',
+    date: '1966-01-31',
+    id: 506775,
+    name: 'Le Combat des chefs',
     volumeId: 74147,
     imageUrl: '4925969-asterix.jpg'
   },
   {
-    date    : '1966-07-31',
-    id      : 506776,
-    name    : 'Astérix chez les Bretons',
+    date: '1966-07-31',
+    id: 506776,
+    name: 'Astérix chez les Bretons',
     volumeId: 74147,
     imageUrl: '4925970-8769082101-97820.jpg'
   },
   {
-    date    : '1966-10-31',
-    id      : 506777,
-    name    : 'Astérix et les Normands',
+    date: '1966-10-31',
+    id: 506777,
+    name: 'Astérix et les Normands',
     volumeId: 74147,
     imageUrl: '4925974-1763181986-816-y.jpg'
   },
   {
-    date    : '1967-07-31',
-    id      : 506778,
-    name    : 'Astérix légionnaire',
+    date: '1967-07-31',
+    id: 506778,
+    name: 'Astérix légionnaire',
     volumeId: 74147,
     imageUrl: '4925975-2167329905-81uqM.jpg'
   },
   {
-    date    : '1968-01-31',
-    id      : 506779,
-    name    : 'Le Bouclier arverne',
+    date: '1968-01-31',
+    id: 506779,
+    name: 'Le Bouclier arverne',
     volumeId: 74147,
     imageUrl: '4925978-eeee.jpg'
   },
   {
-    date    : '1930-05-08',
-    id      : 201735,
-    name    : 'Tintin Au Pays Des Soviets',
+    date: '1930-05-08',
+    id: 201735,
+    name: 'Tintin Au Pays Des Soviets',
     volumeId: 32164,
     imageUrl: '1163607-57206.jpg'
   },
   {
-    date    : '1931-06-11',
-    id      : 201736,
-    name    : 'Tintin Au Congo',
+    date: '1931-06-11',
+    id: 201736,
+    name: 'Tintin Au Congo',
     volumeId: 32164,
     imageUrl: '1163610-57190.jpg'
   },
   {
-    date    : '1932-10-20',
-    id      : 201737,
-    name    : 'Tintin En Amerique',
+    date: '1932-10-20',
+    id: 201737,
+    name: 'Tintin En Amerique',
     volumeId: 32164,
     imageUrl: '1163612-57209.jpg'
   }
