@@ -1,11 +1,10 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 
 @NgModule({
-  declarations: [],
-  imports     : [
+  imports: [
     CommonModule
-  ]
+  ],
+  declarations: []
 })
-export class SharedModule {
-}
+export class SharedModule { }
